[FIX] host파일 삭제
