+++ src/helpers/pages.js
@@ -1,7 +1,7 @@
 import Login from "../pages/Login";
 import KakaoLogin from "../pages/Login/KakaoLogin";
 import Preview from "../pages/Preview";
-import Host from '../pages/Host/Host';
+import Host from "../pages/Host";
 import Member from "../pages/Member/Member";
 import Teacher from "../pages/Login/ExtraInfo/Teacher";
 import Parent from "../pages/Login/ExtraInfo/Parent";
